test(#1024): diff/summary artifacts + lane PASS rows + summary split (f,e,c)
(f) ParityFailure now emits target/cassandra-parity/<lane>.diff and a
    cross-process-merged summary.json (lane -> {status, scenario_id, artifacts})
    on the fail-closed panic; each strict lane records a PASS row on success.
    Added byte/offset-delta/checksum/normalized-JSONL diff formatters in
    parity_support. Workflow already uploads target/cassandra-parity/**.
(e) Every dataset-absent fail-closed path now routes through ParityFailure with
    the manifest scenario id, Cassandra source rule, fixture path, component
    list, and copy-pasteable repro command (index/summary/statistics/data lanes).
(c) Audited sstable_parity_index_db_test, sstabledump_parity_index, and the
    Summary.db strict lane: no placeholder/None-tolerant/unwrap_or-default/
    early-return masking found — the fabricated-reference path was already
    removed under #983. No changes needed; left strict assertions intact.

File-size: split mod strict out of sstabledump_parity_summary.rs (1591->845)
into sstabledump_parity_summary_parts/strict.rs (subdir, not its own test
binary) per campsite rule (#1135).
